typing

+++ spynnaker_integration_tests/test_change_neuron_parameters_between_runs/test_synfire_poisson_if_curr_exp_parameter_test_second_none.py
@@ -18,7 +18,7 @@ from spinnaker_testbase import BaseTestCase
 from spynnaker_integration_tests.scripts import SynfireRunner
 
 n_neurons = 200  # number of neurons in each population
-neurons_per_core = n_neurons / 2
+neurons_per_core = int(n_neurons / 2)
 run_times = [5000, 5000]
 wrap_around = False
 # parameters for population 1 first run
@@ -27,7 +27,7 @@ start_time = 0
 duration = 5000.0
 rate = 2.0
 # parameters for population 2 first run
-set_between_runs = [(1, 'duration', 0)]
+set_between_runs = [(1, 'duration', 0.0)]
 extract_between_runs = False
 record_input_spikes = True
